replaceRevive [nfc]: Move replace-revive logic into its own file.
It would have been better to do this at the start of this series,
for a few reasons [1], but it would be annoying to redo the branch
that way by brute force.

[1] https://github.com/zulip/zulip-mobile/pull/4348#pullrequestreview-555904548
